Change 'login' app to 'user' app
