+++ backend/user/views.py
@@ -1,3 +1,26 @@
-from django.shortcuts import render
+import json
 
-# Create your views here.
+from django.http import HttpResponse
+from django.shortcuts import render, redirect
+from django.contrib.auth import authenticate, login
+from user.forms import UserForm
+
+
+def index(request):
+    return render(request, 'main.html')
+
+def signup(request):
+    if request.method == "POST":
+        form = UserForm(request.POST)
+        if form.is_valid():
+            form.save()
+            username = form.cleaned_data.get('username')
+            password1 = form.cleaned_data.get('password1')
+
+            user = authenticate(username=username, password=password1)
+
+            login(request, user)
+            return redirect('/')
+    else:
+        form = UserForm()
+    return render(request, 'signup.html', {'form' : form})
